Move extract/insert tile ops to language core
